AP-11820: PCA now supports float numbers for inf preservation input
AP-11820 (PCA and PCA Apply should take float numbers as minimum
information fraction)

+++ org.knime.base/src/org/knime/base/node/mine/transformation/pca/apply/PCA2ApplyNodeModel.java
@@ -159,7 +159,7 @@ final class PCA2ApplyNodeModel extends NodeModel {
             dimToReduceTo = m_applySettings.getDimModel().getIntValue();
         } else {
             dimToReduceTo = TransformationUtils.calcDimForGivenInfPreservation(inModel.getTransformationMatrix(),
-                m_applySettings.getInfPreservationModel().getIntValue());
+                m_applySettings.getInfPreservationModel().getDoubleValue());
         }
         return TransformationUtils.createColumnRearranger(dataSpec, inModel.getTransformationMatrix(), dimToReduceTo,
             m_applySettings.getRemoveUsedColsModel().getBooleanValue(), usedColumnNames, transType);
